Add mutational spectrum as input for main.py
Before main.py only accepted a maf annotated with a Ref_Tri column that shows
the trinucleotide context. This allows one to also use as input a tsv of the
base transition counts for each sample, a so called mutational spectrum.
Requirement for the input file is that the column order is the same as the
column transitions in the supplied Stratton_signatures.txt file and the first
column in each row should be the sample name. The values are the number of
times a base transition occurs.

Signed-off-by: Ino de Bruijn <ino@ino.pm>